+++ packages/backend-common/src/util/DockerContainerRunner.ts
@@ -38,6 +38,7 @@ export class DockerContainerRunner implements ContainerRunner {
     mountDirs = {},
     workingDir,
     envVars = {},
+    pullImage = true,
   }: RunContainerOptions) {
     // Show a better error message when Docker is unavailable.
     try {
@@ -48,15 +49,17 @@ export class DockerContainerRunner implements ContainerRunner {
       );
     }
 
-    await new Promise<void>((resolve, reject) => {
-      this.dockerClient.pull(imageName, {}, (err, stream) => {
-        if (err) return reject(err);
-        stream.pipe(logStream, { end: false });
-        stream.on('end', () => resolve());
-        stream.on('error', (error: Error) => reject(error));
-        return undefined;
+    if (pullImage) {
+      await new Promise<void>((resolve, reject) => {
+        this.dockerClient.pull(imageName, {}, (err, stream) => {
+          if (err) return reject(err);
+          stream.pipe(logStream, { end: false });
+          stream.on('end', () => resolve());
+          stream.on('error', (error: Error) => reject(error));
+          return undefined;
+        });
       });
-    });
+    }
 
     const userOptions: UserOptions = {};
     if (process.getuid && process.getgid) {