+++ packages/backend/src/index.ts
@@ -49,6 +49,7 @@ import techdocs from './plugins/techdocs';
 import todo from './plugins/todo';
 import graphql from './plugins/graphql';
 import badges from './plugins/badges';
+import jenkins from './plugins/jenkins';
 import { PluginEnvironment } from './types';
 
 function makeCreateEnv(config: Config) {
@@ -99,6 +100,7 @@ async function main() {
   const kafkaEnv = useHotMemoize(module, () => createEnv('kafka'));
   const graphqlEnv = useHotMemoize(module, () => createEnv('graphql'));
   const badgesEnv = useHotMemoize(module, () => createEnv('badges'));
+  const jenkinsEnv = useHotMemoize(module, () => createEnv('jenkins'));
 
   const apiRouter = Router();
   apiRouter.use('/catalog', await catalog(catalogEnv));
@@ -114,6 +116,7 @@ async function main() {
   apiRouter.use('/proxy', await proxy(proxyEnv));
   apiRouter.use('/graphql', await graphql(graphqlEnv));
   apiRouter.use('/badges', await badges(badgesEnv));
+  apiRouter.use('/jenkins', await jenkins(jenkinsEnv));
   apiRouter.use(notFoundHandler());
 
   const service = createServiceBuilder(module)