+++ packages/catalog-client/src/types/index.ts
@@ -20,5 +20,6 @@ export type {
   CatalogApi,
   CatalogEntitiesRequest,
   CatalogListResponse,
+  CatalogRequestOptions,
 } from './api';
 export { ENTITY_STATUS_CATALOG_PROCESSING_TYPE } from './status';