+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/index.ts
@@ -20,4 +20,7 @@ export * from './debug';
 export * from './fetch';
 export * from './filesystem';
 export * from './publish';
+export * from './github';
+
+export { createFetchCookiecutterAction } from '@backstage/plugin-scaffolder-backend-module-cookiecutter';
 export { runCommand } from './helpers';