+++ plugins/search/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-search
 
+## 0.4.4
+
+### Patch Changes
+
+- 9266b80ab: Adding a type filter to new search
+- Updated dependencies
+  - @backstage/core-components@0.1.6
+  - @backstage/plugin-catalog-react@0.3.1
+
+## 0.4.3
+
+### Patch Changes
+
+- 078d4973e: Handle request errors properly and display them in the results list.
+- Updated dependencies
+  - @backstage/plugin-catalog-react@0.3.0
+
 ## 0.4.2
 
 ### Patch Changes